Added new changes
Copied and pasted query.order_by originally only in the index route to the rest handlers that redirect to the home page.
This is to enable the query to execute no matter what loads the home page.
Removed the finally clauses and db.close because attributes of recent artists and venues raised error.
Apparently closing , expires the session.